Rename mixed_ore_map.lua -> mixed_ore_map_special.lua
