+++ maps/biter_battles_v2/mixed_ore_map_special.lua
