+++ maps/biter_battles_v2/terrain.lua
@@ -2,7 +2,7 @@ local Public = {}
 local LootRaffle = require('functions.loot_raffle')
 local BiterRaffle = require('maps.biter_battles_v2.biter_raffle')
 local bb_config = require('maps.biter_battles_v2.config')
-local mixed_ore_map = require('maps.biter_battles_v2.mixed_ore_map')
+local mixed_ore_map_special = require('maps.biter_battles_v2.mixed_ore_map_special')
 local noise = require('maps.biter_battles_v2.predefined_noise')
 local AiTargets = require('maps.biter_battles_v2.ai_targets')
 local tables = require('maps.biter_battles_v2.tables')
